extract Teacher, Teachers and Room classes from models.py
