+++ backend/room.py
@@ -1,0 +1,31 @@
+from __future__ import annotations
+
+import dataclasses
+
+
+@dataclasses.dataclass
+class Room:
+    house: int | str
+    floor: int | None
+    room_nr: int | None
+    appendix: str = ""
+
+    def to_short(self) -> str:
+        if isinstance(self.house, str):
+            assert self.floor is None
+            return self.house + (str(self.room_nr) if self.room_nr else "") + self.appendix
+
+        if self.floor < 0:
+            return f"-{self.house}{abs(self.floor)}{self.room_nr:02}{self.appendix}"
+        elif self.floor == 0:
+            return f"{self.house}{self.room_nr:02}{self.appendix}"
+        else:
+            return f"{self.house}{self.floor}{self.room_nr:02}{self.appendix}"
+
+    def to_dict(self) -> dict:
+        return {
+            "house": self.house,
+            "floor": self.floor,
+            "room_nr": self.room_nr,
+            "appendix": self.appendix
+        }
