+++ backend/schools/taro.py
@@ -23,10 +23,16 @@ def get_teachers() -> list[Teacher]:
         if teacher_data[1] and teacher_data[2]:
             for name in teacher_images:
                 if name[:-4].startswith(teacher_data[2]) and name[:-4].endswith(teacher_data[1]):
-                    teacher_lst.append(Teacher(abbreviation=teacher_data[0], surname=name[:-4].replace("_", " "), image_path=name))
+                    teacher_lst.append(Teacher(
+                        abbreviation=teacher_data[0],
+                        surname=name[:-4].replace("_", " "),
+                        image_path=name
+                    ))
                     break
             else:
-                teacher_lst.append(Teacher(abbreviation=teacher_data[0], surname=f"{teacher_data[2]} {teacher_data[1]}"))
+                teacher_lst.append(
+                    Teacher(abbreviation=teacher_data[0], surname=f"{teacher_data[2]} {teacher_data[1]}")
+                )
     return teacher_lst
